// src/types/pokemon.ts
export interface PokemonData {
  name: string;
  ability: Abilities[];
  imageUrl: string;
}

export interface Abilities {
  name: string,
  effect: Effect[]
}

export interface Effect {
  effect: string;
  shortEffect: string;
}

export interface PokemonRoot {
  count: number
  next: string
  previous: any
  results: PokemonResult[]
}

export interface PokemonResult {
  name: string
  url: string
}

export interface PokemonDetail {
  abilities: Ability[]
  base_experience: number
  cries: Cries
  forms: Form[]
  game_indices: Index[]
  height: number
  held_items: any[]
  id: number
  is_default: boolean
  location_area_encounters: string
  moves: Mfe[]
  name: string
  order: number
  past_abilities: any[]
  past_types: any[]
  species: Species
  sprites: Sprites
  stats: Stat[]
  types: Type[]
  weight: number
}

export interface Ability {
  ability: Ability2
  is_hidden: boolean
  slot: number
}

export interface Ability2 {
  name: string
  url: string
}

export interface Cries {
  latest: string
  legacy: string
}

export interface Form {
  name: string
  url: string
}

export interface Index {
  game_index: number
  version: Version
}

export interface Version {
  name: string
  url: string
}

export interface Mfe {
  move: Move
  version_group_details: VersionGroupDetail[]
}

export interface Move {
  name: string
  url: string
}

export interface VersionGroupDetail {
  level_learned_at: number
  move_learn_method: MoveLearnMethod
  version_group: VersionGroup
}

export interface MoveLearnMethod {
  name: string
  url: string
}

export interface VersionGroup {
  name: string
  url: string
}

export interface Species {
  name: string
  url: string
}

export interface Sprites {
  back_default: string
  back_female: any
  back_shiny: string
  back_shiny_female: any
  front_default: string
  front_female: any
  front_shiny: string
  front_shiny_female: any
  other: Other
  versions: Versions
}

export interface Other {
  dream_world: DreamWorld
  home: Home
  "official-artwork": OfficialArtwork
  showdown: Showdown
}
// ...
export interface Versions {
  "generation-i": GenerationI
  "generation-ii": GenerationIi
  "generation-iii": GenerationIii
  "generation-iv": GenerationIv
  "generation-v": GenerationV
  "generation-vi": GenerationVi
  "generation-vii": GenerationVii
  "generation-viii": GenerationViii
}
// ...
export interface Stat {
  base_stat: number
  effort: number
  stat: Stat2
}

export interface Stat2 {
  name: string
  url: string
}

export interface Type {
  slot: number
  type: Type2
}

export interface Type2 {
  name: string
  url: string
}


export interface AbilityEffect {
  effect_changes: EffectChange[]
  effect_entries: EffectEntry2[]
  flavor_text_entries: FlavorTextEntry[]
  generation: Generation
  id: number
  is_main_series: boolean
  name: string
  names: Name[]
  pokemon: Pokemon[]
}

export interface EffectChange {
  effect_entries: EffectEntry[]
  version_group: VersionGroup
}

export interface EffectEntry {
  effect: string
  language: Language
}

export interface Language {
  name: string
  url: string
}

export interface VersionGroup {
  name: string
  url: string
}

export interface EffectEntry2 {
  effect: string
  language: Language2
  short_effect: string
}

export interface Language2 {
  name: string
  url: string
}

export interface FlavorTextEntry {
  flavor_text: string
  language: Language3
  version_group: VersionGroup2
}

export interface Language3 {
  name: string
  url: string
}

export interface VersionGroup2 {
  name: string
  url: string
}

export interface Generation {
  name: string
  url: string
}

export interface Name {
  language: Language4
  name: string
}

export interface Language4 {
  name: string
  url: string
}

export interface Pokemon {
  is_hidden: boolean
  pokemon: Pokemon2
  slot: number
}

export interface Pokemon2 {
  name: string
  url: string
}